MLCIL-ICCV2023/requirements.txt
continuum==1.2.3
inplace_abn==1.1.0
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.5.5.64
pandas==1.3.5
Pillow==9.3.0
pycocotools==2.0.4
PyYAML==6.0
randaugment==1.0.2
scipy==1.7.3
termcolor==1.1.0
timm==0.5.4
torch==1.11.0
torchvision==0.12.0
